+++ fiddles/angular2-cli/fiddle-0014-FacebookAuthO/server/config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  AUTH0_DOMAIN: 'AUTH0_DOMAIN',
+  AUTH0_API_AUDIENCE: 'http://localhost:8083/api/',
+  MONGO_URI: process.env.MONGO_URI || 'MONGO_DB_CLOUD_URI'
+};