+++ scripts/bin/angular2-cli/_build.sh
@@ -99,7 +99,7 @@ function build() {
       else
          if [[ -d "dist/$1" ]]; then rm -R "dist/$1"; fi
       fi
-      nvmInstall || exit 1;
+      # nvmInstall || exit 1;
       seederBuild $1 || exit 2;
   )